bond: allow late registration of bonds
In case a bond existed in the systemd during initialization we do not
create a bond internally. As soon as an interface is attached to these
bonds we now allow late bond registration.

Minor bug fixed in the deletion of a lag_id in the controller stub that
is so far not used to create lags on the switch.