nl_l3: adds check for bridge interfaces; adds comments
Adding a termination MAC entry after registering the link local address
has a bug, where the wrong VID is discovered therefore writing the
incorrect VLAN on the TMAC entry on the ASIC.

This PR fixes VID discovery for the bridged ports, thus being able to
write the correct VID. Also featured on this PR is a simplification of
the addition of IP addresses, IPv4 and IPv6, and

Seeing output of "bridge vlan":
```
swbridge         1 PVID Egress Untagged
         15
         4090

port1    1 Egress Untagged
         15 PVID Egress Untagged
```

you would expect `newVlanId = 0x100f (VLAN 15)` set in the switche's
flowtable
```
--  inPort = 1 (Physical)  vlanId:mask = 0x0000:0x1fff (VLAN 0) | GoTo =
20 (Termination MAC) newVlanId = 0x100f (VLAN 15) | priority = 3
   hard_time = 0 idle_time = 0 cookie = 10
```

in some cases it does not set `newVlanId = 0x100f (VLAN 15)` but
`newVlanId = 0x1001 (VLAN 1)` instead

```
--  inPort = 1 (Physical)  vlanId:mask = 0x0000:0x1fff (VLAN 0) | GoTo =
20 (Termination MAC) newVlanId = 0x1001 (VLAN 1) | priority = 3
   hard_time = 0 idle_time = 0 cookie = 19
```

The bug is that this rule was written correctly, but then overwritten by
the re-addition of IP addresses, like the IPv6 link local. Since
baseboxd also configures the VLAN entries on this case, it was
re-writing the correct rule with the wrongly discovered VLAN id.

PR includes a script under bridging/trunk/cntl_trunk_script.sh

Signed-off-by: Rubens Figueiredo <rubens.figueiredo@bisdn.de>