of-dpa: set switch state to up when openflow handshake completes
Openflow spec treats the features request as the handshake, and OF-DPA
does the same. OF-DPA uses sending of the reply to it as the signal to
be ready to send asynchronous messages, so this is essentially means the
switch should be treated as up here.

This fixes a race where OF-DPA sends port_status messages with OFPPR_ADD
before the switch is in state up. These trigger netdev port creation,
but netlink message reception is still off, so the internal mapping
becomes broken.

So we need to make sure we are ready before OF-DPA starts sending these.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>