Revert "tap_man: ensure correct return value enforcement; nbi: prevent changing port status"
On the first link state change after bring up  we will get PORT_ADD
events for all ports, regardless if their state changes.

So the port already existing is an expected case, and we should not
treat it as an error.

This is a partial revert of f917fe5d1dd9a060c87fd6abc3e47edb9afdac8c.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>