nbi_impl: start netlink handling as early as possible
Starting netlink handling by the time the switch goes to STATE_UP is too
late, by that time we might have already received multiple port_status
messages.

Instead just start it on instantiation, like we stop it in destruction.
The additional netlink messages we capture should be harmless, as they
will be for devices we do not care about.

This should really fix the missed netlink messages on connection
establishment, as netlink handling will now be always active.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>