Revert "of-dpa: set switch state to up when openflow handshake completes"
This feature request isn't the completion of the handshake, it's a
separate feature request. The feature request to complete the handshake
is done by rofl-common, and by the time we get the connection it is
already completed.

So this does not do what it is intended to do, and is completely
ineffective.

This reverts commit 9d7d3bfeba6d949454e5ecf77a1b7a10416096b7.