cnetlink: fix ::get_port_id for bridged bond interfaces
The link might be the bridge link interface, and thus not the bond
interface directly, so the type check for bond will fail.

To avoid that, get the AF_UNSPEC link of the same ifindex, which will
lookup the actual interface from the bridge link interface.

This fixes get_port_id returning 0 for bridged bond interfaces.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>