nl_vlan: fix refcount handling for vlans
In add_vlan, only add the vlan entries if the vlan is unknown, else just
increase the refcount.

On vlan remove, we need to check that if the refcount reaches zero and
then delete it, and not just return after decrementing the refcount.
Else we will never remove the vlan because we will keep decrementing the
refcount.

Fixes removing vlan flow entries on deletion of a vlan interface.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>