nl_bridge: relax bridge rtnl_link requirement
If there already is a preexisting bonded interface attached to a bridge,
we might get a early version of the bridge interface that does not has
its kind set. This triggers a null pointer access in the assert.

To work around this, check for kind being set, and in absence, check if
the link's master is itself (which will be for the bridge).

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>