nl_l3: fix unresolved nets prefix comparison
We are looking for the unresolved net that matches the destination, so
we need to check that the comparison returns 0.

Avoids a crash in some cases were we pass a broken address to
nl_route_query. The source of these broken addresses is not yet clear,
but ensuring the prefix matches will ignore them.

Note that we do not need to explicitly check that the family matches as
this is the very first thing nl_addr_cmp{_prefix}() does.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>