nl_l3: add_l3_neigh: fix updating routes waiting on unresolved nexthops
When a route has a unresolved nexthop, we add the route's destination to
a list of nets waiting to be resolved.

When adding a l3 neighbor, we check that list, but the destination of a
neighbor will be the address of it (so the desired nexthop), not the
network waiting for it to be resolved. So we need to store the
unresolved nexthop's address as well.

Instead of expanding the net_resolved_callbacks, there actually already
is a perfectly fine way of doing that by nl_l3 implementing nh_reachable
and registering itself, then adding the route in the callback function.

This way we get all the info, and can remove the net_resolved_callback
completely, as we don't need it anymore.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>