nl_l3: fix memory corruption on {nh,net_}callback removal
It looks like std::deque really does not like it if you try to remove
elements in the middle, and will while removing the element call the
desctructor of the first or last element, whichever is closest.

Since we do route refcounting in the destructors of the *_params, this
causes the wrong routes to drop their references, eventually causing
libnl to free them, and the memory being reused, while we still point at
them.

Fix this by switching to a std::list, which is intended for arbitrary
manipulation.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>