nl_l3: fix nh_reachable callback condition check
rtnl_neigh_get_dst() will be the address of the nexthop, not any network
reachable by that nexthop (which might be more than one), so we need to
compare it with the nexthop address of the nh_params, not the net's
addr.

While at it, drop the superfluous family comparision, since nl_addr_cmp
already does that before comparing the addresses themselves.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>