cnetlink: ignore bridge attachment for untracked interfaces
Our code adding IP addresses and routes has checks for the bridge having
actual port or bond interfaces attached, else they will get ignored.

So we need to make sure that we do not create the bridge too early, and
only do so when the attached interface is one we track, i.e. has a
port_id.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>