Revert "Merge pull request #257 from bisdn/jogo_fix_initial_port_creation_v2"
Interfaces like bridges/ bonds and loopback can receive configuration
before baseboxd has created any interface. There is currently no
explicit check for applying routes/ addresses on these interfaces and
verifying if this config should actually be applied. That has the
consequence of attempting to immediately after the start of baseboxd's netlink thread
to apply flowmods to the ASIC before the connection is alive.

This PR was reversed due to a change in how we process the initial state
of baseboxd. While the PR#257 was aiming to start the netlink
processing as early as possible, a bug was introduced where several
netlink messages will be processed immediately even though the switch
connection is not yet up. Reverting the commit restores the old behavior
to prepare for changed workflow.

This reverts commit 2c852aa25e01cbed27f57d97f785e38664ba2964, reversing
changes made to fb973a7f106a092cb7200e45bef53acc7c9887c3.