cnetlink: handle vrf attach with bridge not set yet
If the VRF attachment happens before any of our interfaces are attached
to the bridge, bridge will be a nullptr so cnetlink::search_fdb will
cause a nullpointer access.

Fix this by guarding the code with a check, and return an empty dequeue
otherwise. If there are no interfaces attached, then it should be
impossible for neighbors (we care about) to exist anyway.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>