nl_bond: don't create bond intefaces on link updates
If we get an update for a bond interface that we do not have created in
the controller, we should just ignore the update.

We will apply the current configuration at the point that we actually
add the bond interface to the controller, once we have one of our port
interfaces attached.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>