nl_l3: add l3 termination MAC and egress group update
When a MAC address on an interface is changed, the L3 configurations
must be changed, specifically the Termination MAC and egress groups.
The MAC addresses on these tables must be updated to reflect the new
state on the interfaces.

cnetlink: enable updates on bridge interfaces
nl_l3: add termination mac adding log

Co-authored-by: Jonas Gorski <jonas.gorski@gmail.com>
Signed-off-by: Rubens Figueiredo <rubens.figueiredo@bisdn.de>