nl_l3: notify routes about IPv6 link local nexthops
While we should not create host entries for IPv6 link local neighbors,
we still may have routes with them as the nexthop.
So we cannot early exit when handling IPv6 link local neighbors, and
still need to go through the nh_(un)reachable_callbacks and notify all
interested parties.

Since we do not add host entries, we do not need to add l3 egress
entries.

This means we might not need to drop a reference, and we cannot use
the return value anymore as an optimization on whether to notify anyone
when a nexthop gets unreachable.

Fixes forwarding of IPv6 routes through IPv6 link local neighbors.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>