nl_bridge: do not set global STP state as VID 1 stp state
If we update the a bridged port's global STP state we update all STGs
with the calculated effective states, and lastly want to set the default
group with STGID 1 to the global state.

But controller::ofdpa_stg_state_port_set takes the VID, not the STGID,
so passing 1 will actually modify VID 1's STP state (only), potentially
undoing a previous block or discard.

So pass it 0, which isn't a valid VID, and which will be resolved
internally to STGID 1.

Fixes: de0c37d ("datamodel: remove ofdpaGlobalStpStatePortSet")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>