nl_bridge: ensure port vlans have a known stp state
Due to message ordering we may ignore the default vlan message, since it
may come before we even claimed the bridge.

We can't process it early, since at this point the bridge object does
not exist, so next best thing fake it when attaching the port (i.e.
old_link is unset).

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>