nl_bridge: dont add fdb entry to neighbors without vlan
Fixes bisdn/basebox/issues/272.
we assume that all fdb entries come assigned to vlan.
When the fdb entry does not have any vlan assigned,
rtnl_neigh_get_vlan() returns -1 because vid is unset,
which we then pass on as vid, which is obviously not a valid VID.

On add/del_neigh_to_fdb, skip adding the entry when the vid is
not valid

Signed-off-by: Rubens Figueiredo <rubens.figueiredo@bisdn.de>
Co-authored-by: Jonas Gorski <jonas.gorski@bisdn.de>