nl_bridge: handle port updates in learn_source_mac
Properly detect and handle macs appearing on different ports than
previously learned, and update the entry instead.

In case a station moves between ports, we need to first find out if the
station was already known at a different port, and then both send a
replace request to the kernel, as well as replace the entry in the
cache.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>