controller: wrap multicast calls
l2_multicast_group_add and l2_multicast_addr_add are always called
together, like l2_multicast_group_remove and l2_multicast_addr_remove
are. So add a helper function for them and use that in nl_bridge, and
mark the other functions private.

Simplifies the handling a bit and abstracts out the specifics how
multicast is implemented in OF-DPA specifically.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>