nl_bridge: fix disabling bridge vlans when there are active mc subscriptions
Disabling bridge vlans may fail if the port is currently subscribed to
multicast groups in that vlan, so make sure we delete all before
attempting to remove the vlan.

The kernel does not flush mdb entries on bridge vlan removal, so they
will only get removed if they time out, which happens way after the vlan
was removed. Unfortunately each membership entry triggers a reference to
the l2 interface group, so removal of the vlan will then fail.

To work around that we need to drop all active multicast memberships
immediately, and in case the vlan is readded before the entries timed
out, restore the entries.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>