nl_bridge: flush cache before removing entries
Avoid a race between flow removed modifications and removing the cache
entries.

Removing the L2 entries from the switch will result in flow removed
notifications. If we are fast enough handling them, we will handle them
before we get to remove the cache entries, the flow removed handler will
try to remote the entries from the kernel. This will fail, since we may
not modify the fdb of a non-existing vlan.

To avoid even trying, we need to remove the entries from the cache
first, so we never try to remove the kernel entries.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>