nl_l3: query cache for the link object
Getting the link object reference from a rtnl_addr object may sometimes
be outdated when there are updates to a link, but no updates to the
address. This creates situations where there is a mismatch between the
old state and current state.

This PR fixes the case where the MAC address in the tap interface is
different than the MAC address in rtnl_addr->a_link.

Fixes https://github.com/bisdn/basebox/issues/362
Co-authored-by: Jonas Gorski <jonas.gorski@bisdn.de>

Signed-off-by: Rubens Figueiredo <rubens.figueiredo@bisdn.de>