add knet manager for using knet based interfaces
Add a port_manager implementation using broadcom KNET interfaces
on switch.

These provide a better performance and lower latency, since packets do
not need to flow through multiple (network) layers anymore, but can only
be used when running on the switch.

We cannot create the interfaces directly by talking with the kernel
module, but need to go through the SDK since we lack required
information. So call OF-DPA's client_drivshell utility directly.

For each KNET interface, we set keeprxtag (so the vlan tag won't be
stripped), and create a filter for forwarding all traffic seen by the
controller to the appropriate interface. The striptag option of filters
is ignored when keeprxtag is enabled for the interface, so we do not
need to disable it.

The interface and filter IDs are assigned linearily, starting with 1 for
interfaces, and 2 for filters, since filter 1 is the default that passes
on packets to the RxAPI of the SDK.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>