nl_l3: replace termination mac refcount with checks for addresses
Replace the refcounts of termination mac entries with simple checks for
the existence of entries on adding addresses, and existing other
addresses on removing other addresses.

This avoids a race on start up where a logical bond or bridge
interface's address may get added/counted twice:

  1. libnl: attach port to bond
  2. libnl: new address on bond
  3. baseboxd: handle attach port event
  4. baseboxd: handle new address event

On attachment, we scan the interface for addresses to add them (since we
only should handle them once there is port interface attached). If an
address gets added before we manage to handle the attachment event, we
will see it in the attachment, and then add it again when handling the
new address event.

So to avoid this, don't try to have valid reference count, and instead
make adding the termination mac entry too often a NOP.

Instead, do a check on address deletion if there are additional
addresses, and only if there are none remove the termination mac
entries.

Since we use VID 1 for untagged traffic by default as well, do a special
check when handling addresses on the untagged or vid 1 tagged
interfaces, and check for addresses on the other one as well (if it exists).

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>