nl_bridge: fix setting initial vlan stp state
When attaching a port to the bridge, the initial state might not
necessarily be still forwarding, the vlan might already be changed to
blocking. So we need to be sure to ask the cache for the current state.

This can especially easily happen for bond interfaces, which we ignore
until a port is attached, but it might already be attached to a bridge
(and have a stp state).

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>