nl_l3: don't treat the dst of on-link routes as unresolved nexthop
On-link routes will have a next hop with just the interface. In this
case rtnl_route_nh_get_via will return a nullptr. In this case, just
ignore the nexthop, and don't add it to unresolved_nexthops.

This then allows dropping the special handling of LINK_SCOPE later on,
since unresolved_nhs won't contain the route itself anymore.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>