port_manager: make destructor virtual
The port_manager is an interface class, and thus functions in this class
should be virtual. If the destructor is not virtual, then the class
implementing port_manager cannot not override the destructor, and the
warning is raised.

```
[3/18] Compiling C++ object baseboxd.p/src_netlink_ctapdev.cc.o
In file included from ../src/netlink/tap_manager.h:13,
from ../src/netlink/ctapdev.cc:18:
../src/netlink/port_manager.h:31:7: warning: 'class
basebox::port_manager' has virtual functions and accessible non-virtual
destructor [-Wnon-virtual-dtor]
31 | class port_manager {
|       ^~~~~~~~~~~~
```

Signed-off-by: Rubens Figueiredo <rubens.figueiredo@bisdn.de>