add a post exit script to remove KNET interfaces
Make sure that all KNET interfaces are removed after baseboxd stopped,
so restarting baseboxd (either intentional or because it crashed) does
not fail due to already existing knet interfaces.

This aligns the behaviour with how it works with the default tap
interfaces, which get removed when baseboxd closes.

The tap interfaces are removed automatically since they are owned by
baseboxd and dropped once baseboxd goes away, but this does not work for
the KNET interfaces, since they are neither owned by baseboxd, nor do
they have any facility to remove them automatically.

The proper way to fix this would be for baseboxd handle sigterm and
gracefully remove them, but the code is currently not equipped well for
that, and will require a larger rework.

As an interim smaller solution, add a python script that directly talks
with the KNET kernel module's ioctl interface to remove all interfaces
and filters, with the exception of the default RxAPI filter, which needs
to be kept.

The IOCTL interface definition is taken from

* https://github.com/Broadcom-Network-Switching-Software/OpenBCM/blob/0b149ddfa3878e65eb217a11dddb999d3e205d03/sdk-6.5.24/include/kcom.h
* https://github.com/Broadcom-Network-Switching-Software/OpenBCM/blob/0b149ddfa3878e65eb217a11dddb999d3e205d03/sdk-6.5.24/systems/linux/kernel/modules/include/bcm-knet.h

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>