nl_l3: properly handle default routes
RTM_GETROUTE does not actually give us the route for a destination net,
it gives us the route to a destination host, with an implicit /32 resp.
/128. This works by accident for all non-default routes (since the first
IP is part of the network), but falls apart for IPv4 default routes,
since 0.0.0.0/32 is the any address, which aliases to localhost.

To fix this, actually ask the netlink cache for a route with the
appropriate destination. As a bonus side effect, the dst of the route
will already have the correct prefix length, and we do not need to
modify it anymore.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>