rofl-util: build_mask_in4: avoid undefined behavior with prefix_len = 0
Calling build_mask_in4 with prefix_len 0 causes a bit shift of 32 in a
32 bit wide variable.

This is undefined behavavior according to e.g. C++11:

> 5.8 Shift operators
>
> The behavior is undefined if the right operand is negative, or greater
> than or equal to the length in bits of the promoted left operand.

The result is that on x86, the shift is done with the modulo of the
variable size, resulting in a 0 bit shift, so this becomes

    ~((1 << (32 - 0)) - 1)
  = ~((1 << (32)) - 1)
  = ~((1 << 0) -1)
  = ~(1 - 1)
  = ~(0)
  = 0xffffffff

which is the opposite of what was intended. Fix this by only doing the
calculation if the prefix_len is positive.

Fixes installation of default routes.

Before:

Table ID 30 (Unicast Routing):   Retrieving all entries. Max entries = 49152, Current entries = 4.
--  etherType = 0x0800 vrf:mask = 0x0000:0x0000 dstIp4 = 0.0.0.0/255.255.255.255 dstIp6 = ::/:: | GoTo = 60 (ACL Policy) groupId = 0x20000001 | priority = 2 hard_time = 0 idle_time = 0 cookie = 13

After:
Table ID 30 (Unicast Routing):   Retrieving all entries. Max entries = 49152, Current entries = 4.
--  etherType = 0x0800 vrf:mask = 0x0000:0x0000 dstIp4 = 0.0.0.0/0.0.0.0 dstIp6 = ::/:: | GoTo = 60 (ACL Policy) groupId = 0x20000001 | priority = 2 hard_time = 0 idle_time = 0 cookie = 13

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>