Copy linter.yml template to enable nable super-linter
Copy the template file [1] from the super-linter repo to activate the
super-linter.

The super-linter is a linter image created and maintained by the GitHub
Services DevOps Engineering team. We can use it to prevent regressions
in clang-format compliance. At the same time, we get an opportunity to
use numerous other linters.

[1] https://github.com/github/super-linter/blob/main/TEMPLATES/linter.yml

Signed-off-by: Roger Luethi <roger.luethi@bisdn.de>