linter.yml: disable failing linters
We added the super-linter to prevent regressions in the clang-format
compliance.

When setting VALIDATE_ALL_CODEBASE to true, we find that a number of
additional linters are triggered and fail. We could just enable
VALIDATE_CLANG_FORMAT, but it may be a good idea of having a record of
all the linters we could use but don't (yet). Also, having linters
enabled by default means we have additional linters active that are
passing already (GITHUB_ACTIONS, GITLEAKS, YAML). Of course there is
always a chance that new, failing linters get added to the linter image
later on.

Signed-off-by: Roger Luethi <roger.luethi@bisdn.de>