tap_manager: set MAC address on created interfaces
Now that we got passed the MAC addresses, we can tell the tap
module to assign them via the ifr_hwaddr.sa_data field.

This uses a (currently) custom extension of the kernel driver. An
unpatched kernel will just ignore it, so this cannot break anything.

Since ctapdev is responsible for creating the tap interfaces, we need to
store the MAC address there as well, and provice an accessor so
tap_manager::recreate_tapdev() can recreate it.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>