nl_vlan: use {add,remove}_ingress_vlan instead of open coding it
Make use of {add,remove}_ingress_vlan helper functions instead of open
coding them.

While this theoretically reduces the amount of log points, the log would
have been the same (as the logs use port_id instead of the member
port_id), and the only reason the call would fail is if the OpenFlow
connection is broken, and in that case it does not matter at which call
it broke.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>