cnetlink: make sure vlan state is restored on detachment from bridge
Make sure we (temporarily) drop the vlans when attaching to a bridge,
and reenable them afterwards when being detached from the bridge.

This is not a proper fix, as this will still break configuration on the
port interface while attached, but it will at least restore the
functionality on detachment.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>

update en/dis calls