nl_bridge: default STP state should be forwarding
When removing a VLAN from a bridged port, we currently set its STP state to
DISABLED, dropping all traffic on it. When the port then gets removed from the
bridge, the VLAN will continue being blocked on this port. To ensure this does
not happen, use FORWARDING as the default state. The VLAN flow/group
memberships will still take care of filtering the traffic while being a bridge
member, so this is fine to do.
This follows a change we did in creation of STG groups, where we switched the
default from DISABLED to FORWARDING.

Co-authored-by: Jonas Gorski <jonas.gorski@bisdn.de>
Signed-off-by: Hilmar Magnusson <hilmar.magnusson@bisdn.de>