cnetlink::link_created(): use is_switch_interface()
When attaching VxLAN interfaces to a bridge, we need to inform the
bridge, and by extension, the VxLAN handling code. With commit
935d0e6fbd9a ("cnetlink: ignore bridge attachment for untracked
interfaces"), this was broken since we started checking for a port_id,
which VxLAN interfaces do not have.

This broke VxLAN setup in various instances, e.g. when the VxLAN
interface was only attached after the access port as added to the
bridge.

To make sure we properly detect VxLAN interfaces, we need to convert the
bridge member link to a base link, as the bridge member link does not
have any of the link type specific attributes like VNI for VxLAN
interfaces.

Fixes: 935d0e6fbd9a ("cnetlink: ignore bridge attachment for untracked interfaces")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>