cnetlink::link_created(): call vxlan::create_endpoint() later
When a VxLAN interface is created, it is not yet attached to a bridge,
so vxlan::create_interface() will fail, but still do some setup.

When the interface is then later attached to a bridge,
vxlan::create_endpoint() may be called a second time, but since the
previous call already did some work, the call will then fail again,
since time because it assumes the previous call succeded.

Since the vxlan::register_bridge() code skips calling
vxlan::create_interface() for VxLAN interfaces not attached to bridges
in its scan of existing interfaces, let's mirror that in
cnetlink::link_created() and don't call vxlan::create_endpoint() there.
It is newly crated, so it cannot be attached to a bridge yet.

But since this only is done once on registration of the bridge, we still
need to call it "manually" for VxLAN interfaces attached at a later
step, so move the vxlan::create_endpoint() to the code handling bridge
attachement.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>