nl_bridge::fdb_timeout(): check ifindex of hit before deletion
nl_cache_search() uses nl_object_identical()[1] to lookup the target
object. This limits the check to unique attributes, which does not
include the ifindex [2].

Therefore nl_cache_search() may return an object for a different port,
in case it moved, which we then may wrongly remove from the cache.

So add a check for the ifindex of the hit before attempting to delete
any neighbors or cache entries.

We could have also alternatively used nl_cache_find() which uses all
attributes, but this function iterates over the whole cache intead of
using the hashtable if possible, so would be much less performant.

[1] https://github.com/thom311/libnl/blob/libnl3_7_0/lib/cache.c#L1113-L1114
[2] https://github.com/thom311/libnl/blob/libnl3_7_0/lib/route/neigh.c#L323

Fixes: 291591d6f313 ("l2 aging added")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>