nl_bridge::add_neigh_to_fdb(): remove old cache item on update
Calling nl_cache_add() for an object cache that supports hashing will
refuse to add an object if it already contains an "identical" object,
where identical is considered having the same master, vid and mac for
AF_BRIDGE neighbors.

Therefore we need to remove the old entry first before we can add an
updated one when we handle neighbors moving between ports.

Without it, we may fail to update the flow if the port moves back again
to its original port, as the outdated entry makes us think we already
have a flow for it.

Fixes: de9b16eea7e4 ("nl_bridge: take over fdb entries instead of creating them")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>