nl_l3: ignore duplicated link routes from FRR
OSPF from FRR 9.0 will sometimes install a duplicate route to an on-link
route, then delete it shortly after:

> Oct 04 08:50:41 accton-as5835-54x baseboxd[2145]: I20231004 08:50:41.874454  2171 cnetlink.cc:1223] route_route_apply: new route (nl_obj=0x7fa654022370) inet 10.0.0.0/24 table main type unicast
> Oct 04 08:50:41 accton-as5835-54x baseboxd[2145]:     preferred-src 10.0.0.1 scope link priority 0 protocol kernel
> Oct 04 08:50:41 accton-as5835-54x baseboxd[2145]:     nexthop dev port54 <>
> ...
> Oct 04 08:50:42 accton-as5835-54x baseboxd[2145]: I20231004 08:50:42.178130  2171 cnetlink.cc:1223] route_route_apply: new route (nl_obj=0x7fa654022c30) inet 10.0.0.0/24 table main type unicast
> Oct 04 08:50:42 accton-as5835-54x baseboxd[2145]:     scope global priority 0x14 protocol 0xbc
> Oct 04 08:50:42 accton-as5835-54x baseboxd[2145]:     nexthop dev port54 <>
> ...
> Oct 04 08:50:42 accton-as5835-54x baseboxd[2145]: I20231004 08:50:42.188859  2171 cnetlink.cc:1258] route_route_apply: del route (nl_obj=0x7fa654022c30) inet 10.0.0.0/24 table main type unicast
> Oct 04 08:50:42 accton-as5835-54x baseboxd[2145]:     scope global priority 0x14 protocol 0xbc
> Oct 04 08:50:42 accton-as5835-54x baseboxd[2145]:     nexthop dev port54 <>

This causes baseboxd to first (try to) install the route a second time,
then delete it when deleting the duplicate route, landing in a state
where the route is missing from the flows.

To work around this, check routes if they have an associated protocol
that isn't kernel, and a priority higher than zero, and if so, if we
already have a route for that destination. If so, ignore the route.

This is far from proper handling of priorities/metrics, but allows us to
avoid a broken state due to duplicate routes.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>