cnetlink: just ignore bridges with unsupported configuration
Ignore bridges with unsupported configuration instead of treating it as
fatal state and crashing. Do this by keeping a list of ignored bridges,
and ignoring them for attach/detatchment until the bridges get deleted.

This avoid baseboxd getting in a broken state, as we won't apply any
configuration from these bridges.

This avoids getting in an invalid OF-DPA state requiring a reboot to
recover, and potentially going into a crash loop.

Note: this does not cover the case of disabling vlan_aware e.g. via
sysfs after attaching ports, but that is likely a rare scenario.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>