nl_vxlan::create_endpoint(): fix error log on add neigh failure
There is no link variable in this nl_vxlan::create_endpoint(), only
vxlan_link and br_link. This was hidden due to the OBJ_CAST(), but when
the OBJ_CAST() was dropped, g++ started to complain.

Fix this by replacing link with neigh, since neigh is what the text
before the variable says.

Fixes the following warning:

../git/src/netlink/nl_vxlan.cc: In member function 'int basebox::nl_vxlan::create_endpoint(rtnl_link*, rtnl_link*, nl_addr*)':
../git/src/netlink/nl_vxlan.cc:674:45: warning: the address of 'int link(const char*, const char*)' will never be NULL [-Waddress]
  674 |                  << ") to add l2 neigh " << link;
      |                                             ^~~~

Fixes: ec04fe2d229e ("vxlan: add existing neighbors for new endpoint")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>