of-dpa: reset switch state on connect
To ensure that the baseboxd and switch state are in sync, reset the
switch state on connect:

* delete all flows
* delete all groups
* delete all tunnels and their tenents and nexthops
* delete all bond ports
* reset STG groups to default (all vlans in group 1)

This ensures that the we have a valid state if baseboxd was restarted,
either voluntarily (due to a restart) or involuntarily (due to a crash).

Since this may interrupt some workflows, e.g. if you set up custom ACL
rules before startup of baseboxd, make it possible to disable this
behaviour via a flag.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>