nl_bridge: fix removing aged out neighbors from fdb
When fixing moving neighbors between ports we removed the ifindex from
the filter neigh to find a neighbor at any port with the mac address.

But since we use the same neigh to construct the fdb delete message,
this caused the message to miss the ifindex, and the kernel was
rejecting the deletion. This resulting in us not updating out local
cache of known layer 2 neighbors, and failing to relearn them.

Fix this by just using the found entry from our local cache, n_lookup.

Fixes: a566d4caa1b0 ("nl_bridge::fdb_timeout(): check ifindex of hit before deletion")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>