nl_l3: keep IPv6LL route enabled if any interface still has it
Currently we unconditionally delete the IPv6LL route when it get's
deleted from the routing table. But this is a per-interface route, so
just because one route is removed does not mean that there are none
left.

So the following sequence

    $ ip link set up port1
    $ ip link set up port2
    $ ip link set down port1

will result in the IPv6LL route being disabled, although port1 still has
one.

So add a check on deletion if there are any remaining interfaces left
with the IPv6LL route, and only actually delete it if there aren't any.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>